configure: Avoid warnings if av_always_inline is disabled
This macro is disabled if --enable-small or --disable-optimizations are set.
Currently, this leads to warnings about functions being defined but not used,
for functions in header files.

By defining av_always_inline as inline or as av_unused, we avoid these
warnings.

This doesn't make a normal build with --enable-small any larger,
since the compiler probably chooses not to inline these functions even
if they're marked as inline.


git-svn-id: file:///var/local/repositories/ffmpeg/trunk@26032 9553f0bf-9b14-0410-a0b8-cfaf0461ba5b