Add mapping mechanism for reordered_opaque values.
The CrystalHD 'timeStamp' value is the only tool we have to
match an encoded frame to a decoded one. However, libcrystalhd
will mangle these values if they are not proper timestamps.

ffmpeg uses arbitrary 'reordered_opaque' values to order frames,
so we need a way to map them without being vulnerable to mangling.

This change adds a queue-like structure that allows us to pass
a fake timestamp to libcrystalhd and get back to the real opaque
value after decoding.