Try harder to get the second field of an interlaced frame.
Sometimes, the hardware will output each field on a separate ProcOutput
call, even when the source compressed fields were submitted together. If
we don't try really hard to receive both output fields in one decode()
iteration, we'll cause the pipeline to grow and potentially exhaust the
input buffer, so keep trying to get that second field.