Peek the next frame/field to detect actual interlacing.
The hardware will set the interlaced flag on MBAFF content
that is actually a field pair. This makes it impossible to
separate from PAFF content without being to check if the next
ouput is a second field or another frame.

We can peek the picture number and if it is the same, then there
is a second field and it's really interlaced. If not, it's a
mislabeled MBAFF field pair.