Handle PAFF content.
Ok, so there are three forms that I've seen interlaced content take.

1) MBAFF as a single field pair. Treat it as a progressive frame and
all is well.

2) MPEG2 where a single input avpkt results in two output fields. I
call ProcOutput twice for each input to keep things balanced.

3) PAFF where each avpkt contains one field. In this case, I have to
return after getting the first field to keep the 1:1 input:output
ratio.

Note that we still can't detect the difference between PAFF and MBAFF.
My previously committed approach of peeking the timestamp doesn't work
because the next field might not be sufficiently decoded for its picture
number to be available - and in the case of 1080i, this is usually the
case. Looks like we need a firmware fix.