Switch to using existing ffmpeg code for mp4 -> annex b conversion.
There is existing code to do the extradata conversion in the mp4 to
annex b bitstream filter. It needs a bit of a kludge to work around
mplayer running the decoder twice because it probes the file first.

So we have to copy the extradata aside and put it back so that the
next time round we know it's really mp4 and not annex b.