Use ffmpeg's internal routines for copying frames.
We can use av_image_copy_plane to do the copy, and av_image_get_linesize
to calculate the byte width of the frame - removing the need to hardcode
it. Of course, all the code assumes a packed image format; a planar one
would require separate handling.

Also took the opportunity to add 70012 detection and do the source stride
calculations that are needed for it.