Add adaptive delay mechanism.
The amount of time to wait at the end of a decode() call varies depending
on the format being decoded, so a fixed wait value will not always work.
As such, I'm introducing a mechanism that increases the delay if we see
decode calls failing due to no frames being ready, but only after the
initial pipeline is established.

Note that this mechanism is imperfect for at least two reasons:

1) If the pipeline length legitimately changes mid-stream (eg: switching
from progressive to interlaced content), we will increase the delay
when it's likely that no increase was needed.

2) There is a tension between giving the decoder enough time to work
and procesing the video data fast enough to display in realtime. This
is especially evident when dealing with 1080i60 content. Some samples
do show the delay increasing too much and decode rate falling below
realtime.

Nevertheless, the overall behaviour is a net improvement.